refactor: dto 수정
home과 team/plan에서 dto를 공유하기 위한 수정
