+++ src/main/java/com/kuit/conet/dto/request/plan/PlanRequest.java
@@ -1,4 +1,4 @@
-package com.kuit.conet.dto.request.home;
+package com.kuit.conet.dto.request.plan;
 
 import lombok.*;
 
@@ -6,7 +6,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class HomePlanRequest {
+public class PlanRequest {
     String searchDate;
     // 특정 달의 조회: "yyyy-MM"
     // 특정 날짜 조회: "yyyy-MM-dd"
