+++ src/main/java/com/kuit/conet/dto/response/plan/DayPlanResponse.java
@@ -1,4 +1,4 @@
-package com.kuit.conet.dto.response.home;
+package com.kuit.conet.dto.response.plan;
 
 import com.kuit.conet.domain.FixedPlan;
 import lombok.*;
@@ -10,7 +10,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class HomeDayPlanResponse {
+public class DayPlanResponse {
     int count;
     List<FixedPlan> plans;
 }
