+++ src/main/java/com/kuit/conet/service/HomeService.java
@@ -3,10 +3,10 @@ package com.kuit.conet.service;
 import com.kuit.conet.dao.HomeDao;
 import com.kuit.conet.domain.FixedPlan;
 import com.kuit.conet.domain.WaitingPlan;
-import com.kuit.conet.dto.request.home.HomePlanRequest;
-import com.kuit.conet.dto.response.home.HomeDayPlanResponse;
-import com.kuit.conet.dto.response.home.HomeMonthPlanResponse;
-import com.kuit.conet.dto.response.home.HomeWaitingPlanResponse;
+import com.kuit.conet.dto.request.plan.PlanRequest;
+import com.kuit.conet.dto.response.plan.DayPlanResponse;
+import com.kuit.conet.dto.response.plan.MonthPlanResponse;
+import com.kuit.conet.dto.response.plan.WaitingPlanResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,35 +21,35 @@ import java.util.List;
 public class HomeService {
     private final HomeDao homeDao;
 
-    public HomeMonthPlanResponse getPlanOnMonth(HttpServletRequest httpRequest, HomePlanRequest planRequest) {
+    public MonthPlanResponse getPlanInMonth(HttpServletRequest httpRequest, PlanRequest planRequest) {
         List<Integer> planDates = new ArrayList<>();
 
         Long userId = Long.parseLong((String) httpRequest.getAttribute("userId"));
         String searchDate = planRequest.getSearchDate(); // yyyy-MM
 
-        List<String> dateList = homeDao.getPlanOnMonth(userId, searchDate);
+        List<String> dateList = homeDao.getPlanInMonth(userId, searchDate);
         for(String tempDate : dateList) {
             Integer date = Integer.parseInt(tempDate.split("-")[2]);
             planDates.add(date);
         }
 
-        return new HomeMonthPlanResponse(planDates.size(), planDates);
+        return new MonthPlanResponse(planDates.size(), planDates);
     }
 
-    public HomeDayPlanResponse getPlanOnDay(HttpServletRequest httpRequest, HomePlanRequest planRequest) {
+    public DayPlanResponse getPlanOnDay(HttpServletRequest httpRequest, PlanRequest planRequest) {
         Long userId = Long.parseLong((String) httpRequest.getAttribute("userId"));
         String searchDate = planRequest.getSearchDate(); // yyyy-MM-dd
 
         List<FixedPlan> plans = homeDao.getPlanOnDay(userId, searchDate);
 
-        return new HomeDayPlanResponse(plans.size(), plans);
+        return new DayPlanResponse(plans.size(), plans);
     }
 
-    public HomeWaitingPlanResponse getWaitingPlanOnDay(HttpServletRequest httpRequest) {
+    public WaitingPlanResponse getWaitingPlanOnDay(HttpServletRequest httpRequest) {
         Long userId = Long.parseLong((String) httpRequest.getAttribute("userId"));
 
         List<WaitingPlan> plans = homeDao.getWaitingPlanOnDay(userId);
 
-        return new HomeWaitingPlanResponse(plans.size(), plans);
+        return new WaitingPlanResponse(plans.size(), plans);
     }
 }
